Improve InvoiceIntegrationClient interface
Move Espoo specific code into its own files
